backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bcrypt<4.0
pydantic[email]
sqlmodel
alembic
python-dotenv
aiosqlite
httpx-oauth
